+++ packages/delir-core/src/Engine/AssetProxy.ts
