+++ packages/delir-core/src/Engine/IRenderingStreamObserver.ts
