+++ packages/delir-core/src/Engine/renderer/RendererBase.ts
