+++ packages/delir-core/src/index.ts
@@ -1,17 +1,17 @@
-import * as Engine from './engine'
+import * as Engine from './Engine'
 import * as Exceptions from './exceptions'
 import * as PluginSupport from './plugin-support'
 import * as Project from './project/index'
 import * as Values from './Values'
 
-import PreRenderRequest from './engine/PreRenderingRequest'
-import RenderRequest from './engine/RenderRequest'
+import PreRenderRequest from './Engine/PreRenderingRequest'
+import RenderRequest from './Engine/RenderRequest'
 import PluginBase from './plugin-support/plugin-base'
 import PluginRegistry from './plugin-support/plugin-registry'
 import PostEffectBase from './plugin-support/PostEffectBase'
 import Type, { AnyParameterTypeDescriptor, TypeDescriptor } from './plugin-support/type-descriptor'
 
-import * as KeyframeCalcurator from './engine/KeyframeCalcurator'
+import * as KeyframeCalcurator from './Engine/KeyframeCalcurator'
 import * as ProjectHelper from './helper/project-helper'
 import ProjectMigrator from './helper/ProjectMigrator'
