+++ packages/delir-core/src/plugin-support/type-descriptor.ts
@@ -1,4 +1,4 @@
-import AssetProxy from '../engine/pipeline/AssetProxy'
+import AssetProxy from '../Engine/pipeline/AssetProxy'
 import PluginLoadFailException from '../exceptions/plugin-load-fail-exception'
 import defaults from '../helper/defaults'
